Split __init__ file into multiple and add test
