+++ dev-requirements.txt
@@ -1,2 +1,3 @@
 pytest==8.0.0
 responses==0.24.1
+mock==5.1.0
